make semantics depend on painting instead of rendering
