+++ packages/flutter/lib/src/rendering/proxy_box.dart
@@ -13,7 +13,6 @@ import 'package:flutter/services.dart';
 import 'package:vector_math/vector_math_64.dart';
 
 import 'box.dart';
-import 'debug.dart';
 import 'layer.dart';
 import 'object.dart';
